+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+- BasicFrame: don't desync the subsurface from the main one. This avoids
+  graphical glitches where the borders are not drawn exactly the same size
+  as the contents.
+
 ## 0.2.1 -- 2018-05-03
 
 - Add `DoubleMemPool` for double buffering, and use it to